+++ src/main/java/univ/kgu/carely/domain/meet/service/impl/MeetingServiceImpl.java
@@ -10,7 +10,7 @@ import univ.kgu.carely.domain.meet.entity.Meeting;
 import univ.kgu.carely.domain.meet.entity.MeetingStatus;
 import univ.kgu.carely.domain.meet.entity.Memory;
 import univ.kgu.carely.domain.meet.repository.MeetingRepository;
-import univ.kgu.carely.domain.meet.repository.MemoryRepository;
+import univ.kgu.carely.domain.meet.repository.memory.MemoryRepository;
 import univ.kgu.carely.domain.meet.service.MeetingService;
 import univ.kgu.carely.domain.member.entity.Member;
 import univ.kgu.carely.domain.member.repository.MemberRepository;
